Unpack feature files from within TCK jar
Previously we downloaded feature files directly from GitHub, which
worked well until the TCK dev tools expanded. Now we need a tight fit
between TCK dev tools and feature files, and so openCypher bundle the latter
inside the JAR downloaded via Maven, and we unpack them here.
