+++ community/cypher/compatibility-suite/pom.xml
@@ -89,7 +89,7 @@
     <dependency>
       <groupId>org.opencypher</groupId>
       <artifactId>tck</artifactId>
-      <version>1.2016-04-05</version>
+      <version>1.2016-04-11</version>
     </dependency>
 
     <!-- JSON -->
